+++ packages/components/modules/comments/native/CommentItem/index.tsx
@@ -1,18 +1,9 @@
-import { FC, useCallback, useRef } from 'react'
+import { FC } from 'react'
 
 import { AvatarWithPlaceholder } from '@baseapp-frontend/design-system/components/native/avatars'
-import { BottomDrawer } from '@baseapp-frontend/design-system/components/native/drawers'
-import {
-  EditIcon,
-  PinIcon,
-  ShareIcon,
-  TrashIcon,
-} from '@baseapp-frontend/design-system/components/native/icons'
 import { Text } from '@baseapp-frontend/design-system/components/native/typographies'
 import { View } from '@baseapp-frontend/design-system/components/native/views'
-import { useTheme } from '@baseapp-frontend/design-system/providers/native'
 
-import { type BottomSheetModal } from '@gorhom/bottom-sheet'
 import { Pressable } from 'react-native'
 import { useRefetchableFragment } from 'react-relay'
 
@@ -27,7 +18,7 @@ import { CommentItemProps } from './types'
 
 const CommentItem: FC<CommentItemProps> = ({
   comment: commentRef,
-  onEdit,
+  onLongPress,
   Timestamp = DefaultTimestamp,
   CommentReplyButton = DefaultCommentReplyButton,
   CommentReactionButton = DefaultCommentReactionButton,
@@ -37,124 +28,54 @@ const CommentItem: FC<CommentItemProps> = ({
     commentRef,
   )
 
-  const theme = useTheme()
   const styles = createStyles()
-  const bottomDrawerRef = useRef<BottomSheetModal | undefined>(undefined)
 
   const replyToComment = () => {
     // TODO (in another story)
     console.log('Reply button clicked')
   }
 
-  const handleLongPress = useCallback(() => {
-    bottomDrawerRef.current?.present()
-  }, [])
-
-  const handleSheetChanges = useCallback((index: number) => {
-    if (index === -1) {
-      // Sheet is closed
-    }
-  }, [])
-
-  const handleMenuAction = useCallback(
-    (_action: string) => {
-      bottomDrawerRef.current?.close()
-      if (_action === 'edit') {
-        onEdit?.(comment)
-      }
-    },
-    [comment, onEdit],
-  )
-
   if (!comment) {
     return null
   }
 
   return (
-    <>
-      <Pressable onLongPress={handleLongPress}>
-        <View style={styles.rootContainer}>
-          <View style={styles.avatarContainer}>
-            <AvatarWithPlaceholder imgSource={comment.user?.avatar?.url} />
-          </View>
-          <View style={styles.bodyContainer}>
-            <Text variant="subtitle2" color="high">
-              {comment.user?.fullName}
-            </Text>
-            <Text variant="body2" color="high">
-              {comment.body}
-              {comment.isEdited ? (
-                <Text variant="body2" color="disabled" style={{ fontSize: 12 }}>
-                  {' '}
-                  (edited)
-                </Text>
-              ) : (
-                ''
-              )}
-            </Text>
-            <View style={styles.footerContainer}>
-              <View style={styles.buttonContainer}>
-                <CommentReactionButton target={comment} shouldUseBottomSheetSafeComponents />
-                <CommentReplyButton
-                  onReply={replyToComment}
-                  commentId={comment.id}
-                  shouldUseBottomSheetSafeComponents
-                />
-              </View>
-              <View style={styles.timestampContainer}>
-                <Timestamp date={comment.created} />
-              </View>
+    <Pressable onLongPress={() => onLongPress?.(comment)}>
+      <View style={styles.rootContainer}>
+        <View style={styles.avatarContainer}>
+          <AvatarWithPlaceholder imgSource={comment.user?.avatar?.url} />
+        </View>
+        <View style={styles.bodyContainer}>
+          <Text variant="subtitle2" color="high">
+            {comment.user?.fullName}
+          </Text>
+          <Text variant="body2" color="high">
+            {comment.body}
+            {comment.isEdited ? (
+              <Text variant="body2" color="disabled" style={{ fontSize: 12 }}>
+                {' '}
+                (edited)
+              </Text>
+            ) : (
+              ''
+            )}
+          </Text>
+          <View style={styles.footerContainer}>
+            <View style={styles.buttonContainer}>
+              <CommentReactionButton target={comment} shouldUseBottomSheetSafeComponents />
+              <CommentReplyButton
+                onReply={replyToComment}
+                commentId={comment.id}
+                shouldUseBottomSheetSafeComponents
+              />
+            </View>
+            <View style={styles.timestampContainer}>
+              <Timestamp date={comment.created} />
             </View>
           </View>
         </View>
-      </Pressable>
-      <BottomDrawer
-        bottomDrawerRef={bottomDrawerRef}
-        handleSheetChanges={handleSheetChanges}
-        snapPoints={['30%']}
-      >
-        <View style={styles.bottomDrawerActionContainer}>
-          <Pressable onPress={() => handleMenuAction('share')} style={styles.bottomDrawerPressable}>
-            <ShareIcon width={20} height={20} color={theme.colors.object.high} />
-            <Text variant="body2" color="high">
-              Share Comment
-            </Text>
-          </Pressable>
-          {comment.canPin && (
-            <Pressable onPress={() => handleMenuAction('pin')} style={styles.bottomDrawerPressable}>
-              <PinIcon width={20} height={20} color={theme.colors.object.high} />
-              <Text variant="body2" color="high">
-                Pin Comment
-              </Text>
-            </Pressable>
-          )}
-          {comment.canChange && (
-            <Pressable
-              onPress={() => handleMenuAction('edit')}
-              style={styles.bottomDrawerPressable}
-            >
-              <EditIcon width={20} height={20} color={theme.colors.object.high} />
-              <Text variant="body2" color="high">
-                Edit Comment
-              </Text>
-            </Pressable>
-          )}
-        </View>
-        {comment.canDelete && (
-          <View style={styles.bottomDrawerDivider}>
-            <Pressable
-              onPress={() => handleMenuAction('delete')}
-              style={styles.bottomDrawerPressable}
-            >
-              <TrashIcon width={20} height={20} color={theme.colors.error.main} />
-              <Text variant="body2" style={{ color: theme.colors.error.main }}>
-                Delete Comment
-              </Text>
-            </Pressable>
-          </View>
-        )}
-      </BottomDrawer>
-    </>
+      </View>
+    </Pressable>
   )
 }
