+++ packages/components/modules/comments/native/CommentsList/styles.ts
@@ -5,4 +5,24 @@ export const createStyles = () =>
     listContainer: {
       backgroundColor: 'transparent',
     },
+    bottomDrawerActionContainer: {
+      paddingHorizontal: 16,
+      paddingVertical: 12,
+      gap: 4,
+      borderBottomWidth: 1,
+      borderBottomColor: 'surface.disabled',
+    },
+    bottomDrawerDivider: {
+      paddingHorizontal: 16,
+      paddingVertical: 12,
+      gap: 4,
+      borderBottomWidth: 0,
+    },
+    bottomDrawerPressable: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'flex-start',
+      gap: 8,
+      padding: 8,
+    },
   })
